clean [major]: renaming operations to operation and canonicalization to canonicalize. added docs/examples.md
