+++ docs/examples.md
@@ -1,0 +1,1 @@
+# Examples Index
