+++ docs/technical/reference/backend.md
@@ -1,0 +1,3 @@
+# Backend
+
+::: midstack.backend
