+++ docs/technical/reference/operations-interface.md
@@ -50,4 +50,4 @@ classDiagram
   }
 ```
 
-::: midstack.interface.analog.operations
+::: midstack.interface.analog.operation
