+++ mkdocs.yml
@@ -26,6 +26,7 @@ nav:
   - Tutorials:
       - Analog: tutorials/analog.md
   - Examples:
+      - Index: examples.md
       - Analog:
           - Rabi Flopping: examples/one_qubit_rabi_flopping.ipynb
           - Adiabatic state:
