+++ src/midstack/compiler/analog/passes/canonicalize.py
@@ -1,7 +1,7 @@
 from midstack.compiler.rewriter import Chain, FixedPoint
 from midstack.compiler.walk import Post, Pre, In
 from midstack.compiler.analog.rewrite.canonicalize import *
-from midstack.compiler.analog.verify.canonicalization import *
+from midstack.compiler.analog.verify.canonicalize import *
 from midstack.compiler.analog.verify.operator import VerifyHilberSpaceDim
 from midstack.compiler.math.rules import (
     DistributeMathExpr,
