+++ src/midstack/compiler/analog/verify/canonicalize.py
