+++ src/midstack/interface/analog/__init__.py
@@ -1,2 +1,2 @@
 from .operator import *
-from .operations import *
+from .operation import *
