+++ src/midstack/interface/analog/operation.py
